lod: .lrtc files and a streaming pool that loads what the view wants most
writeLrtc/readLrtc: page-aligned blocks in the device's packing, so a chunk
maps alone and reading is a copy. StreamingPool keeps the merged levels on
the device and loads chunks on worker threads into a fixed store, most wanted
first -- the cut returns each chunk's largest wanting cell edge -- taking free
slots, then chunks not wanted, then chunks wanted less than half as much.
lrt convert writes .lrtc; lrt render/bench read it whole or stream it with
--stream-budget. On train_30k the PLY, the .lrtc and a 4-of-16-slot stream
write byte-identical images.

+++ modules/lod/include/lrt/lod/Lod.h
@@ -107,9 +107,11 @@ struct CutStats {
     uint32_t splats = 0;     ///< the cloud's own splats drawn
     uint32_t merged = 0;     ///< merged Gaussians drawn
     uint32_t available = 0;  ///< splats in the cloud
-    /// Streamed clouds: per chunk, 1 when this view wants its splats (on the
-    /// device or not). Empty for a cloud that is all in memory.
-    std::vector<uint8_t> needs;
+    /// Streamed clouds: per chunk, 0 when this view does not want its splats,
+    /// else how much -- the largest projected edge, in 1/16 pixels, of the
+    /// cells wanting them -- on the device or not. Empty for a cloud that is
+    /// all in memory.
+    std::vector<uint32_t> needs;
 };
 
 class CutSelector {
